test: update trainer test
