+++ tests/jax/integration/mock_systems.py
@@ -1,0 +1,200 @@
+import functools
+from datetime import datetime
+from typing import Any
+import optax
+import launchpad as lp
+
+from mava.systems.jax import System, ippo
+from mava.utils.environments import debugging_utils
+from mava.utils.loggers import logger_utils
+
+
+def mock_system_single_process() -> System:
+    """Mock of ippo system in the single process case"""
+    # Environment.
+    environment_factory = functools.partial(
+        debugging_utils.make_environment,
+        env_name="simple_spread",
+        action_space="discrete",
+    )
+
+    # Networks.
+    def network_factory(*args: Any, **kwargs: Any) -> Any:
+        return ippo.make_default_networks(  # type: ignore
+            policy_layer_sizes=(32, 32),
+            critic_layer_sizes=(64, 64),
+            *args,
+            **kwargs,
+        )
+
+    # Checkpointer appends "Checkpoints" to checkpoint_dir.
+    base_dir = "~/mava"
+    mava_id = str(datetime.now())
+    checkpoint_subpath = f"{base_dir}/{mava_id}"
+    # Log every [log_every] seconds.
+    log_every = 1
+    logger_factory = functools.partial(
+        logger_utils.make_logger,
+        directory=base_dir,
+        to_terminal=True,
+        to_tensorboard=True,
+        time_stamp=mava_id,
+        time_delta=log_every,
+    )
+    # Optimizer.
+    optimizer = optax.chain(
+        optax.clip_by_global_norm(40.0),
+        optax.adam(1e-4),
+    )
+
+    # Create ippo system
+    test_system = ippo.IPPOSystem()
+
+    # Build the test_system
+    test_system.build(
+        environment_factory=environment_factory,
+        network_factory=network_factory,
+        logger_factory=logger_factory,
+        experiment_path=checkpoint_subpath,
+        optimizer=optimizer,
+        executor_parameter_update_period=10,
+        multi_process=False,  # Single process case
+        run_evaluator=True,
+        num_executors=1,
+        use_next_extras=False,
+        max_queue_size=5000,
+        sample_batch_size=2,
+        max_in_flight_samples_per_worker=4,
+        num_workers_per_iterator=-1,
+        rate_limiter_timeout_ms=-1,
+        nodes_on_gpu=[],
+        sequence_length=4,
+        period=4,
+    )
+
+    return test_system
+
+
+def mock_system_multi_thread() -> System:
+    """Mock of ippo system in the multi thread case"""
+    # Environment.
+    environment_factory = functools.partial(
+        debugging_utils.make_environment,
+        env_name="simple_spread",
+        action_space="discrete",
+    )
+
+    # Networks.
+    def network_factory(*args: Any, **kwargs: Any) -> Any:
+        return ippo.make_default_networks(  # type: ignore
+            policy_layer_sizes=(32, 32),
+            critic_layer_sizes=(64, 64),
+            *args,
+            **kwargs,
+        )
+
+    # Checkpointer appends "Checkpoints" to checkpoint_dir.
+    base_dir = "~/mava"
+    mava_id = str(datetime.now())
+    checkpoint_subpath = f"{base_dir}/{mava_id}"
+
+    # Log every [log_every] seconds.
+    log_every = 10
+    logger_factory = functools.partial(
+        logger_utils.make_logger,
+        directory=base_dir,
+        to_terminal=True,
+        to_tensorboard=True,
+        time_stamp=mava_id,
+        time_delta=log_every,
+    )
+
+    # Optimizer.
+    optimizer = optax.chain(
+        optax.clip_by_global_norm(40.0),
+        optax.adam(1e-4),
+    )
+
+    # Create ippo system
+    test_system = ippo.IPPOSystem()
+
+    # Build the system
+    test_system.build(
+        environment_factory=environment_factory,
+        network_factory=network_factory,
+        logger_factory=logger_factory,
+        experiment_path=checkpoint_subpath,
+        optimizer=optimizer,
+        executor_parameter_update_period=1,
+        multi_process=True,
+        run_evaluator=True,
+        num_executors=1,
+        max_queue_size=500,
+        use_next_extras=False,
+        sample_batch_size=5,
+        nodes_on_gpu=[],
+        lp_launch_type=lp.LaunchType.TEST_MULTI_THREADING,
+    )
+    return test_system
+
+
+def mock_system_multi_process() -> System:
+    """Mock of ippo system in the multi process case"""
+    # Environment.
+    environment_factory = functools.partial(
+        debugging_utils.make_environment,
+        env_name="simple_spread",
+        action_space="discrete",
+    )
+
+    # Networks.
+    def network_factory(*args: Any, **kwargs: Any) -> Any:
+        return ippo.make_default_networks(  # type: ignore
+            policy_layer_sizes=(254, 254, 254),
+            critic_layer_sizes=(512, 512, 256),
+            *args,
+            **kwargs,
+        )
+
+    # Checkpointer appends "Checkpoints" to checkpoint_dir.
+    base_dir = "~/mava"
+    mava_id = str(datetime.now())
+    checkpoint_subpath = f"{base_dir}/{mava_id}"
+
+    # Log every [log_every] seconds.
+    log_every = 10
+    logger_factory = functools.partial(
+        logger_utils.make_logger,
+        directory=base_dir,
+        to_terminal=True,
+        to_tensorboard=True,
+        time_stamp=mava_id,
+        time_delta=log_every,
+    )
+
+    # Optimizer.
+    optimizer = optax.chain(
+        optax.clip_by_global_norm(40.0), optax.scale_by_adam(), optax.scale(-1e-4)
+    )
+
+    # Create the ippo system
+    test_system = ippo.IPPOSystem()
+
+    # Build the system.
+    test_system.build(
+        environment_factory=environment_factory,
+        network_factory=network_factory,
+        logger_factory=logger_factory,
+        experiment_path=checkpoint_subpath,
+        optimizer=optimizer,
+        executor_parameter_update_period=1,
+        multi_process=True,
+        run_evaluator=True,
+        num_executors=1,
+        max_queue_size=500,
+        use_next_extras=False,
+        sample_batch_size=5,
+        nodes_on_gpu=[],
+        lp_launch_type=lp.LaunchType.LOCAL_MULTI_PROCESSING,
+    )
+    return test_system
